Extract OsString utilities to their own moule in artichoke_backend
