+++ artichoke-backend/src/ffi.rs
@@ -2,18 +2,17 @@
 //!
 //! These functions are unsafe. Use them carefully.
 
-use bstr::{ByteSlice, ByteVec};
 use std::borrow::Cow;
 use std::error;
-use std::ffi::{OsStr, OsString};
 use std::fmt;
 use std::mem;
 use std::ptr::{self, NonNull};
 
+use spinoso_exception::Fatal;
+
 use crate::class_registry::ClassRegistry;
 use crate::core::ConvertMut;
 use crate::error::{Error, RubyException};
-use crate::extn::core::exception::{ArgumentError, Fatal};
 use crate::state::State;
 use crate::sys;
 use crate::Artichoke;
@@ -123,96 +122,6 @@ impl From<InterpreterExtractError> for Box<dyn RubyException> {
 
 impl From<Box<InterpreterExtractError>> for Box<dyn RubyException> {
     fn from(exception: Box<InterpreterExtractError>) -> Box<dyn RubyException> {
-        exception
-    }
-}
-
-/// Convert a byte slice to a platform-specific [`OsStr`].
-///
-/// Unsupported platforms fallback to converting through `str`.
-#[inline]
-pub fn bytes_to_os_str(value: &[u8]) -> Result<&OsStr, ConvertBytesError> {
-    value.to_os_str().map_err(|_| ConvertBytesError::new())
-}
-
-/// Convert a platform-specific [`OsStr`] to a byte slice.
-///
-/// Unsupported platforms fallback to converting through `str`.
-#[inline]
-pub fn os_str_to_bytes(value: &OsStr) -> Result<&[u8], ConvertBytesError> {
-    <[u8]>::from_os_str(value).ok_or_else(ConvertBytesError::new)
-}
-
-/// Convert a platform-specific [`OsString`] to a byte vec.
-///
-/// Unsupported platforms fallback to converting through `String`.
-#[inline]
-pub fn os_string_to_bytes(value: OsString) -> Result<Vec<u8>, ConvertBytesError> {
-    Vec::from_os_string(value).map_err(|_| ConvertBytesError::new())
-}
-
-#[derive(Default, Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub struct ConvertBytesError {
-    _private: (),
-}
-
-impl ConvertBytesError {
-    /// Constructs a new, default `ConvertBytesError`.
-    #[must_use]
-    pub const fn new() -> Self {
-        Self { _private: () }
-    }
-}
-
-impl fmt::Display for ConvertBytesError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str("Could not convert between bytes and platform string")
-    }
-}
-
-impl error::Error for ConvertBytesError {}
-
-impl RubyException for ConvertBytesError {
-    fn message(&self) -> Cow<'_, [u8]> {
-        Cow::Borrowed(b"invalid byte sequence")
-    }
-
-    fn name(&self) -> Cow<'_, str> {
-        "ArgumentError".into()
-    }
-
-    fn vm_backtrace(&self, interp: &mut Artichoke) -> Option<Vec<Vec<u8>>> {
-        let _ = interp;
-        None
-    }
-
-    fn as_mrb_value(&self, interp: &mut Artichoke) -> Option<sys::mrb_value> {
-        let message = interp.convert_mut(self.message());
-        let value = interp.new_instance::<ArgumentError>(&[message]).ok().flatten()?;
-        Some(value.inner())
-    }
-}
-
-impl From<ConvertBytesError> for Error {
-    fn from(exception: ConvertBytesError) -> Self {
-        Self::from(Box::<dyn RubyException>::from(exception))
-    }
-}
-
-impl From<Box<ConvertBytesError>> for Error {
-    fn from(exception: Box<ConvertBytesError>) -> Self {
-        Self::from(Box::<dyn RubyException>::from(exception))
-    }
-}
-
-impl From<ConvertBytesError> for Box<dyn RubyException> {
-    fn from(exception: ConvertBytesError) -> Box<dyn RubyException> {
-        Box::new(exception)
-    }
-}
-
-impl From<Box<ConvertBytesError>> for Box<dyn RubyException> {
-    fn from(exception: Box<ConvertBytesError>) -> Box<dyn RubyException> {
         exception
     }
 }
